+++ src/generate/gen_base.cpp
@@ -530,7 +530,7 @@ void BaseCodeGenerator::CollectEventHandlers(Node* node, std::vector<NodeEvent*>
                 if (node->getParent()->isGen(gen_wxContextMenuEvent))
                     m_ctx_menu_events.push_back(&iter.second);
                 else
-                    m_events.push_back(&iter.second);
+                    events.push_back(&iter.second);
             }
         }
     }